requirements.txt
flask
murf-python-sdk
deepgram-sdk==3.*
google-generativeai
python-dotenv
SpeechRecognition
pyaudio